Fix recapthca widget deserialization return.  Refs #2156.
This only broke when upgrading various SFARI libraries, we're guessing
it's the deform/colander upgrade.  We're also guessing it only worked
before because of an unexposed bug.  Since the return from the recaptcha
widget is ignored it may have been just casting the dict this widget
returned into a string which was ignored and now we're guessing
deform/colander is being stricter about the return.  So we now just
return the validity message from recaptcha.